Fix dangling pointer references in clip_view.
Refactor clip_view -> clip_ref_t and clip_view_t.
Use project_t to find out if track_t pointer is valid, then use track_t to find if clip_t pointer is valid.
Handle CMD_CREATE_VIEW and CMD_RESET_UI_DEFAULT_LAYOUT in DawCtrl instead of DawInstance

+++ src/gui/properties/properties_table.cpp
@@ -667,7 +667,7 @@ void addCommonPropertiesFromGui(SafeRef<guibase>& ref, guibase& gui, std::vector
 #ifdef TRACK_ALLOCATIONS_GUIBASE
     rows.push_back({{tblstr{"allocId"}, tblint{gui.allocId}}});
 #endif
-    rows.push_back({{tblstr{"refId"}, tblint{gui.safeRef.refId}}});
+    rows.push_back({{tblstr{"refId"}, tblint{int64_t(gui.safeRef.refId)}}});
     String strTheme = gui.theme->name+StringFormat("[%7zX]", reinterpret_cast<uint64_t>(gui.theme));
     rows.push_back({{tblstr{"theme"}, tblString{strTheme, 1}}});
     rows.push_back({{tblstr{"label"}, tblString{gui.label, 1}}});
